services/dsc_api/requirements.txt
Django>=4.2
djangorestframework
python-dotenv
pytest
pytest-django
django-cors-headers
pandas
scipy
matplotlib
plotly